refactor to use lib file to enable calling from the multi-call binary
